Added animation while waiting for CLI to clean

+++ src/main/java/org/example/Cleaner.java
@@ -1,16 +1,21 @@
 package org.example;
 
 import java.io.File;
-import java.util.Arrays;
-import java.util.Comparator;
-import java.util.PriorityQueue;
-import java.util.Queue;
+import java.util.*;
+
 import static org.example.Utils.isNumericVersion;
 
 public class Cleaner {
 
     private final String BASE_DIRECTORY;
     private final Comparator<File> COMPARATOR;
+
+    private static final String[] FILES_INSIDE_VERSION_FOLDER = {
+            "_remote.repositories",
+            ".pom",
+            ".pom.sha1",
+            ".jar"
+    };
 
     public Cleaner(String BASE_DIRECTORY, int comparatorIndex,boolean reversed) {
         this.BASE_DIRECTORY = BASE_DIRECTORY;
@@ -22,7 +27,6 @@ public class Cleaner {
 
     public void clean(){
         var baseFile = new File(BASE_DIRECTORY);
-
         traverse(baseFile);
     }
     private boolean traverse(File dir){
@@ -30,26 +34,34 @@ public class Cleaner {
         File[] inFiles = dir.listFiles();
         boolean isVersionsFolder = false;
 
-        assert inFiles != null;
+        if (inFiles == null || inFiles.length == 0) {
+            return false;
+        }
 
         for(File file : inFiles){
             if(file.isDirectory()){
+                // If it returns true it means we are in the versions folder
                 if(traverse(file)) {
                     isVersionsFolder = true;
                     break;
                 }
             }
-            if(file.getName().endsWith(".jar")){
-                return true;
+            for(String inNames : FILES_INSIDE_VERSION_FOLDER){
+                if(file.getName().endsWith(inNames)){
+                    return true;
+                }
             }
         }
+        // If we are in a versions folder
         if (isVersionsFolder){
+
             Queue<File> pq = new PriorityQueue<>(COMPARATOR);
 
             Arrays.stream(inFiles)
                     .filter((e) -> isNumericVersion(e.getName()))
                     .forEach(pq::offer);
 
+            // Only the one with the highest priority remains
             var survivor = pq.poll();
 
             while(!pq.isEmpty()){
@@ -63,12 +75,19 @@ public class Cleaner {
 
         return false;
     }
+
+    /**
+     * Recursively all files inside the given directory and the directory
+     * itself
+     * @param directory
+     */
     private boolean deleteDirectory(File directory) {
         if (directory.isDirectory()) {
             File[] files = directory.listFiles();
             if (files != null) {
                 for (File file : files) {
-                    deleteDirectory(file);
+                    if(file.isDirectory()) deleteDirectory(file);
+                    else file.delete();
                 }
             }
         }
